+++ src/BlazorWebFormsComponents/AdRotator.razor.cs
@@ -62,25 +62,7 @@ namespace BlazorWebFormsComponents
 		public Unit Width { get; set; }
 
 		[Parameter]
-		public bool Font_Bold { get; set; }
-
-		[Parameter]
-		public bool Font_Italic { get; set; }
-
-		[Parameter]
-		public string Font_Names { get; set; }
-
-		[Parameter]
-		public bool Font_Overline { get; set; }
-
-		[Parameter]
-		public FontUnit Font_Size { get; set; }
-
-		[Parameter]
-		public bool Font_Strikeout { get; set; }
-
-		[Parameter]
-		public bool Font_Underline { get; set; }
+		public FontInfo Font { get; set; } = new FontInfo();
 
 		private string CalculatedStyle => this.ToStyle().Build().NullIfEmpty();
