+++ src/BlazorWebFormsComponents/Button.razor.cs
@@ -6,7 +6,6 @@ namespace BlazorWebFormsComponents
 {
 	public partial class Button : ButtonBaseComponent, IHasStyle
 	{
-
 		[Parameter] public WebColor BackColor { get; set; }
 		[Parameter] public WebColor BorderColor { get; set; }
 		[Parameter] public BorderStyle BorderStyle { get; set; }
@@ -15,13 +14,7 @@ namespace BlazorWebFormsComponents
 		[Parameter] public WebColor ForeColor { get; set; }
 		[Parameter] public Unit Height { get; set; }
 		[Parameter] public Unit Width { get; set; }
-		[Parameter] public bool Font_Bold { get; set; }
-		[Parameter] public bool Font_Italic { get; set; }
-		[Parameter] public string Font_Names { get; set; }
-		[Parameter] public bool Font_Overline { get; set; }
-		[Parameter] public FontUnit Font_Size { get; set; }
-		[Parameter] public bool Font_Strikeout { get; set; }
-		[Parameter] public bool Font_Underline { get; set; }
+		[Parameter] public FontInfo Font { get; set; } = new FontInfo();
 
 		private string CalculatedStyle => this.ToStyle().Build().NullIfEmpty();
