+++ src/BlazorWebFormsComponents/Style/IFontStyle.cs
@@ -1,0 +1,7 @@
+﻿namespace BlazorWebFormsComponents
+{
+	public interface IFontStyle
+	{
+		FontInfo Font { get; set; }
+	}
+}
