+++ src/BlazorWebFormsComponents/Style/IHasStyle.cs
@@ -1,0 +1,7 @@
+﻿namespace BlazorWebFormsComponents
+{
+	public interface IHasStyle : IHasLayoutStyle, IFontStyle
+	{
+		
+	}
+}
