+++ src/BlazorWebFormsComponents/Style/IHasTableItemStyle.cs
@@ -1,0 +1,7 @@
+﻿namespace BlazorWebFormsComponents
+{
+	public interface IHasTableItemStyle : IHasLayoutTableItemStyle, IFontStyle
+	{
+
+	}
+}
