+++ src/BlazorWebFormsComponents/Validations/BaseValidator.razor.cs
@@ -29,13 +29,7 @@ namespace BlazorWebFormsComponents.Validations
 		[Parameter] public HorizontalAlign HorizontalAlign { get; set; }
 		[Parameter] public VerticalAlign VerticalAlign { get; set; }
 		[Parameter] public Unit Width { get; set; }
-		[Parameter] public bool Font_Bold { get; set; }
-		[Parameter] public bool Font_Italic { get; set; }
-		[Parameter] public string Font_Names { get; set; }
-		[Parameter] public bool Font_Overline { get; set; }
-		[Parameter] public FontUnit Font_Size { get; set; }
-		[Parameter] public bool Font_Strikeout { get; set; }
-		[Parameter] public bool Font_Underline { get; set; }
+		[Parameter] public FontInfo Font { get; set; } = new FontInfo();
 
 		/// <summary>
 		/// Override all style properties if it's not null
